requirements.txt
discord
discord-py-slash-command
pyTelegramBotAPI
pyyaml
boto3
requests